content of file.cpp moved into libtiledb.cpp
